CRUD Product management, success/error modal, fixed issue with images not saving to the public disk correctly

+++ database/migrations/2021_09_14_163622_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('title');
             $table->string('slug')->unique();
             $table->text('description');
-            $table->double('price', 8, 2);
+            $table->integer('price');
             $table->string('image')->nullable();
             $table->string('file')->nullable();
             $table->timestamps();
